Modify the name of header file
