+++ include/kernel.h
@@ -6,7 +6,4 @@
 
 
 
-
-
-
 #endif
