--- include/my_stddef.h
@@ -1,10 +1,0 @@
-#ifndef MY_STDDEF_H
-#define MY_STDDEF_H
-
-
-
-
-#define NULL ((void*)0)
-
-
-#endif
