+++ include/stddef.h
@@ -1,0 +1,10 @@
+#ifndef STDDEF_H
+#define STDDEF_H
+
+
+
+
+#define NULL ((void*)0)
+
+
+#endif
