+++ include/stdint.h
@@ -1,5 +1,5 @@
-#ifndef MY_STDINT_H
-#define MY_STDINT_H
+#ifndef STDINT_H
+#define STDINT_H
 
 #ifndef __bool_defined
 typedef int                          bool;
