+++ src/string.c
